+++ src/auto_survey/__init__.py
@@ -5,7 +5,7 @@ import logging
 from dotenv import load_dotenv
 from termcolor import colored
 
-fmt = f"[%(asctime)s]\n{colored('%(message)s', 'light_yellow')}\n"
-logging.basicConfig(level=logging.INFO, format=fmt, datefmt="%Y-%m-%d %H:%M:%S")
+fmt = colored("%(message)s", "light_yellow") + "\n"
+logging.basicConfig(level=logging.INFO, format=fmt)
 
 load_dotenv(dotenv_path=".env")