+++ src/auto_survey/cli.py
@@ -9,6 +9,7 @@ from pathlib import Path
 
 import click
 import litellm
+from termcolor import colored
 from tqdm.auto import tqdm
 
 from auto_survey.data_models import LiteLLMConfig
@@ -120,7 +121,11 @@ def main(
 
     # Summarise each paper
     for paper in tqdm(
-        iterable=papers, desc="Summarising papers", unit="paper", ascii="—▰"
+        iterable=papers,
+        desc=colored("Summarising papers", "light_yellow"),
+        unit="paper",
+        ascii="—▰",
+        colour="yellow",
     ):
         paper.summary = summarise_paper(
             paper=paper, topic=topic, verbose=verbose, litellm_config=litellm_config
@@ -133,7 +138,7 @@ def main(
         for paper in papers
         if is_relevant_paper(paper=paper, topic=topic, litellm_config=litellm_config)
     ]
-    logger.info(
+    logger.debug(
         f"After reading the full papers, {len(papers):,} papers continue to be "
         "relevant to the topic."
     )
@@ -143,16 +148,59 @@ def main(
         topic=topic, relevant_papers=papers, litellm_config=litellm_config
     )
     markdown_path.write_text(literature_survey)
-    logger.info(f"Wrote Markdown literature survey to {markdown_path.as_posix()}")
+    logger.info(f"Markdown: {markdown_path.as_posix()}")
 
-    # Try to convert the Markdown to PDF using Pandoc
-    subprocess.run(
-        ["pandoc", "--from=markdown", "--to=pdf", f"--output={pdf_path}"],
-        input=literature_survey,
-        encoding="utf-8",
-        check=True,
+    # Convert the Markdown to PDF using Pandoc
+    pandoc_installed = (
+        subprocess.run(
+            ["pandoc", "--version"], capture_output=True, text=True
+        ).returncode
+        == 0
     )
-    logger.info(f"Wrote PDF literature survey to {pdf_path.as_posix()}")
+    pdflatex_installed = (
+        subprocess.run(
+            ["pdflatex", "--version"], capture_output=True, text=True
+        ).returncode
+        == 0
+    )
+    if not pandoc_installed and not pdflatex_installed:
+        logger.warning(
+            "We cannot convert the Markdown to PDF because neither Pandoc nor "
+            "pdflatex is installed. Please install both of these and try again. "
+            "Pandoc installation can be found at https://pandoc.org/installing.html "
+            "and PDFLatex installation can be found at "
+            "https://www.latex-project.org/get/. When these are installed, you can "
+            f"convert the Markdown at {markdown_path.as_posix()} to PDF by running "
+            f"`pandoc --from=markdown --to=pdf --output={pdf_path.as_posix()} "
+            f"{markdown_path.as_posix()}` in your terminal."
+        )
+    elif not pandoc_installed:
+        logger.warning(
+            "We cannot convert the Markdown to PDF because Pandoc is not installed. "
+            "Please install Pandoc and try again. Installation instructions can be "
+            "found at https://pandoc.org/installing.html. When Pandoc is installed, "
+            f"you can convert the Markdown at {markdown_path.as_posix()} to PDF by "
+            f"running `pandoc --from=markdown --to=pdf --output={pdf_path.as_posix()} "
+            f"{markdown_path.as_posix()}` in your terminal."
+        )
+    elif not pdflatex_installed:
+        logger.warning(
+            "We cannot convert the Markdown to PDF because pdflatex is not "
+            "installed. Please install pdflatex and try again. Installation "
+            "instructions can be found at https://www.latex-project.org/get/. When "
+            "pdflatex is installed, you can convert the Markdown at "
+            f"{markdown_path.as_posix()} to PDF by running "
+            f"`pandoc --from=markdown --to=pdf --output={pdf_path.as_posix()} "
+            f"{markdown_path.as_posix()}` in your terminal."
+        )
+    else:
+        subprocess.run(
+            ["pandoc", "--from=markdown", "--to=pdf", f"--output={pdf_path}"],
+            input=literature_survey,
+            encoding="utf-8",
+            check=True,
+        )
+        logger.info(f"PDF: {pdf_path.as_posix()}")
 
 
 if __name__ == "__main__":